Fix timedemo in qw.
td_lastframe needs to be a float in qw. It seems the meaning changed from
nq to qw, but that's rather bogus. I need to look further into juat what's
going on, but at least now timedemo works.

+++ qw/include/client.h
@@ -163,7 +163,7 @@ typedef struct {
 	int         prevtime;
 	double      basetime;
 	qboolean    timedemo;
-	int         td_lastframe;		// to meter out one message a frame
+	float       td_lastframe;		// to meter out one message a frame
 	int         td_startframe;		// host_framecount at start
 	float       td_starttime;		// realtime at second frame of timedemo
 } client_static_t;
